+++ app/views/admin/profile.view.php
@@ -54,7 +54,7 @@
 <?php require ('partials/footer.php')?>
 
     <script src="<?php echo url()?>public/js/libs/jquery-ui.js"></script>
-    <script src="<?php echo url()?>public/js/custom.js"></script>
+    <script src="<?php echo url()?>public/js/app.bundle.js"></script>
 <script>
     $(document).on('submit', '#change_password', function(e){
         e.preventDefault();
