+++ npm-debug.log
@@ -1,0 +1,48 @@
+0 info it worked if it ends with ok
+1 verbose cli [ 'C:\\Program Files\\nodejs\\node.exe',
+1 verbose cli   'C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js',
+1 verbose cli   'run',
+1 verbose cli   'dev' ]
+2 info using npm@4.1.2
+3 info using node@v7.7.2
+4 verbose run-script [ 'predev', 'dev', 'postdev' ]
+5 info lifecycle acc@1.0.0~predev: acc@1.0.0
+6 silly lifecycle acc@1.0.0~predev: no script for predev, continuing
+7 info lifecycle acc@1.0.0~dev: acc@1.0.0
+8 verbose lifecycle acc@1.0.0~dev: unsafe-perm in lifecycle true
+9 verbose lifecycle acc@1.0.0~dev: PATH: C:\Users\vlado\AppData\Roaming\nvm\v7.7.2\node_modules\npm\bin\node-gyp-bin;D:\xampp\htdocs\acc\node_modules\.bin;D:\chrome\downloads\cmder_mini\bin;D:\chrome\downloads\cmder_mini\vendor\bin;D:\chrome\downloads\cmder_mini\vendor\conemu-maximus5\ConEmu\Scripts;D:\chrome\downloads\cmder_mini\vendor\conemu-maximus5;D:\chrome\downloads\cmder_mini\vendor\conemu-maximus5\ConEmu;C:\Program Files (x86)\Common Files\Oracle\Java\javapath;C:\Windows\system32;C:\Windows;C:\Windows\System32\Wbem;C:\Windows\System32\WindowsPowerShell\v1.0\;C:\Windows\System32\OpenSSH\;C:\Users\vlado\AppData\Roaming\nvm;C:\Program Files\nodejs;C:\Program Files\Git\cmd;C:\Program Files (x86)\OpenSSH\bin;D:\xampp\php;C:\ProgramData\ComposerSetup\bin;C:\HashiCorp\Vagrant\bin;C:\Program Files (x86)\NVIDIA Corporation\PhysX\Common;C:\Users\vlado\AppData\Local\Programs\Python\Python35\Scripts\;C:\Users\vlado\AppData\Local\Programs\Python\Python35\;C:\Users\vlado\AppData\Local\Microsoft\WindowsApps;C:\Users\vlado\AppData\Local\Programs\Microsoft VS Code\bin;C:\Users\vlado\AppData\Roaming\npm;C:\Users\vlado\AppData\Roaming\nvm;C:\Program Files\nodejs;C:\Users\vlado\AppData\Roaming\Composer\vendor\bin;C:\Program Files\Git\mingw64\bin;C:\Program Files\Git\usr\bin;D:\chrome\downloads\cmder_mini
+10 verbose lifecycle acc@1.0.0~dev: CWD: D:\xampp\htdocs\acc
+11 silly lifecycle acc@1.0.0~dev: Args: [ '/d /s /c', 'webpack --mode development' ]
+12 silly lifecycle acc@1.0.0~dev: Returned: code: 2  signal: null
+13 info lifecycle acc@1.0.0~dev: Failed to exec dev script
+14 verbose stack Error: acc@1.0.0 dev: `webpack --mode development`
+14 verbose stack Exit status 2
+14 verbose stack     at EventEmitter.<anonymous> (C:\Users\vlado\AppData\Roaming\nvm\v7.7.2\node_modules\npm\lib\utils\lifecycle.js:279:16)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at EventEmitter.emit (events.js:194:7)
+14 verbose stack     at ChildProcess.<anonymous> (C:\Users\vlado\AppData\Roaming\nvm\v7.7.2\node_modules\npm\lib\utils\spawn.js:40:14)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at ChildProcess.emit (events.js:194:7)
+14 verbose stack     at maybeClose (internal/child_process.js:899:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
+15 verbose pkgid acc@1.0.0
+16 verbose cwd D:\xampp\htdocs\acc
+17 error Windows_NT 10.0.19041
+18 error argv "C:\\Program Files\\nodejs\\node.exe" "C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js" "run" "dev"
+19 error node v7.7.2
+20 error npm  v4.1.2
+21 error code ELIFECYCLE
+22 error acc@1.0.0 dev: `webpack --mode development`
+22 error Exit status 2
+23 error Failed at the acc@1.0.0 dev script 'webpack --mode development'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the acc package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     webpack --mode development
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs acc
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls acc
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
