+++ src/js/main.js
@@ -1,0 +1,4 @@
+'use strict'
+import custom from './custom';
+custom();
+
