+++ src/js/test.js
@@ -1,0 +1,1 @@
+//alert('Hello');
